pyproject.toml
[project]
name = "brainwave3d"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.0.0,<3.0.0",
    "fastapi>=0.110.0,<0.120.0",
    "motor>=3.3.0,<4.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "pydantic>=2.4.0,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "pymongo>=4.6.0,<5.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "uvicorn[standard]>=0.27.0,<0.29.0",
]
